# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar-Save Backend API
  description: |
    REST API for the Stellar-Save decentralised rotational savings platform.

    ## Authentication
    Most endpoints are public. Endpoints under `/api/user/*` require a Bearer JWT
    obtained from `POST /api/auth/verify`. Pass it as:
    ```
    Authorization: Bearer <token>
    ```

    ## Versioning
    The API is versioned via URL path prefix:
    - `/api/v1/` — stable, production-ready
    - `/api/v2/` — current development version (some routes return 501)

    Unversioned legacy paths (`/health`, `/recommendations`, etc.) are deprecated
    and will be removed on 2027-01-01.

    ## Rate Limiting
    - General endpoints: 100 req / 15 min per IP
    - Auth & admin endpoints: 10 req / 15 min per IP
    - Analytics write endpoints: stricter limits apply (see individual routes)
  version: 1.0.0
  contact:
    name: Stellar-Save
    url: https://github.com/Xoulomon/Stellar-Save
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.stellar-save.app
    description: Production
...
paths:
...
  /api/v1/notifications/preferences/{userId}:
    get:
      tags: [Notifications]
      summary: Get notification preferences for a user
      operationId: notifPrefsGet
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferences'
    put:
      tags: [Notifications]
      summary: Update notification preferences
      operationId: notifPrefsUpdate
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationPreferencesUpdate'
      responses:
        '200':
          description: Preferences updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  preferences:
                    $ref: '#/components/schemas/NotificationPreferences'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/notifications/device-token:
    post:
      tags: [Notifications]
      summary: Register a device token for push notifications
      operationId: notifDeviceTokenRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, deviceToken, platform]
              properties:
                userId:
                  type: string
                deviceToken:
                  type: string
                platform:
                  type: string
                  enum: [ios, android, web]
      responses:
        '201':
          description: Device token registered
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/notifications/device-token/{userId}/{deviceToken}:
    delete:
      tags: [Notifications]
      summary: Unregister a device token
      operationId: notifDeviceTokenDelete
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: deviceToken
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Device token unregistered
  /api/v1/notifications/send-email:
    post:
      tags: [Notifications]
      summary: Send an email notification
      operationId: notifSendEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, templateId, templateData]
              properties:
                userId:
                  type: string
                to:
                  type: string
                  format: email
                templateId:
                  type: string
                templateData:
                  type: object
                subject:
                  type: string
      responses:
        '202':
          description: Email sent
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: User has disabled this notification type
  /api/v1/notifications/send-push:
    post:
      tags: [Notifications]
      summary: Send a push notification
      operationId: notifSendPush
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deviceToken, templateId, templateData]
              properties:
                userId:
                  type: string
                deviceToken:
                  type: string
                templateId:
                  type: string
                templateData:
                  type: object
                title:
                  type: string
                body:
                  type: string
      responses:
        '202':
          description: Push notification sent
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/notifications/queue:
    post:
      tags: [Notifications]
      summary: Queue a notification for later delivery
      operationId: notifQueue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, templateKey, recipient, templateData, notificationType]
              properties:
                userId:
                  type: string
                templateKey:
                  type: string
                recipient:
                  type: string
                templateData:
                  type: object
                notificationType:
                  type: string
                  enum: [email, push, web_push]
                priority:
                  type: integer
                  default: 0
                scheduledFor:
                  type: string
                  format: date-time
      responses:
        '202':
          description: Notification queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  queueId:
                    type: string
  /api/v1/notifications/process-queue:
    post:
      tags: [Notifications]
      summary: Process queued notifications
      operationId: notifProcessQueue
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                batchSize:
                  type: integer
                  default: 100
      responses:
        '200':
          description: Queue processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  processedCount:
                    type: integer
  /api/v1/notifications/history/{userId}:
    get:
      tags: [Notifications]
      summary: Get notification history for a user
      operationId: notifHistory
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Notification history
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  count:
                    type: integer
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/NotificationRecord'
...
  /api/v1/notifications/templates:
    get:
      tags: [Notifications]
      summary: List all active notification templates
      operationId: notifTemplatesList
      responses:
        '200':
          description: Templates list
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  templates:
                    type: array
                    items:
                      $ref: '#/components/schemas/NotificationTemplate'
    post:
      tags: [Notifications]
      summary: Create a notification template (admin only)
      operationId: notifTemplatesCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationTemplateInput'
      responses:
        '201':
          description: Template created
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/notifications/templates/{templateKey}:
    get:
      tags: [Notifications]
      summary: Get a specific notification template
      operationId: notifTemplatesGet
      parameters:
        - name: templateKey
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationTemplate'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Notifications]
      summary: Update a notification template (admin only)
      operationId: notifTemplatesUpdate
      parameters:
        - name: templateKey
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subject:
                  type: string
                htmlContent:
                  type: string
                textContent:
                  type: string
                active:
                  type: boolean
      responses:
        '200':
          description: Template updated
  /api/v1/notifications/vapid-public-key:
    get:
      tags: [Notifications]
      summary: Get VAPID public key for web push subscriptions
      operationId: notifVapidKey
      responses:
        '200':
          description: VAPID public key
          content:
            application/json:
              schema:
                type: object
                properties:
                  publicKey:
                    type: string
        '503':
          description: Web push not configured
  /api/v1/notifications/subscribe:
    post:
      tags: [Notifications]
      summary: Register a browser push subscription
      operationId: notifSubscribe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, subscription]
              properties:
                userId:
                  type: string
                subscription:
                  $ref: '#/components/schemas/WebPushSubscription'
      responses:
        '201':
          description: Subscription registered
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          description: Web push not configured
    delete:
      tags: [Notifications]
      summary: Remove a browser push subscription
      operationId: notifUnsubscribe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint]
              properties:
                endpoint:
                  type: string
                  format: uri
      responses:
        '200':
          description: Subscription removed
  /api/v1/notifications/unsubscribe/{token}:
    post:
      tags: [Notifications]
      summary: One-click unsubscribe via token (email link)
      operationId: notifOneClickUnsubscribe
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Unsubscribed (returns HTML page)
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Invalid token
  /api/v1/notifications/resubscribe/{userId}:
    post:
      tags: [Notifications]
      summary: Re-subscribe a user to notifications
      operationId: notifResubscribe
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User resubscribed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by POST /api/auth/verify
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    NotificationPreferences:
      type: object
      properties:
        userId:
          type: string
        emailNotifications:
          type: boolean
        pushNotifications:
          type: boolean
        contributionReminders:
          type: boolean
        groupUpdates:
          type: boolean
        payoutNotifications:
          type: boolean
        emailFrequency:
          type: string
          enum: [immediate, daily, weekly, never]
    NotificationPreferencesUpdate:
      type: object
      properties:
        emailNotifications:
          type: boolean
        pushNotifications:
          type: boolean
        contributionReminders:
          type: boolean
        groupUpdates:
          type: boolean
        payoutNotifications:
          type: boolean
        emailFrequency:
          type: string
          enum: [immediate, daily, weekly, never]
    NotificationRecord:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
        templateKey:
          type: string
        status:
          type: string
          enum: [queued, sent, failed]
        sentAt:
          type: string
          format: date-time
    NotificationTemplate:
      type: object
      properties:
        templateKey:
          type: string
        templateName:
          type: string
        templateType:
          type: string
          enum: [email, push, web_push]
        subject:
          type: string
        htmlContent:
          type: string
        textContent:
          type: string
        placeholders:
          type: array
          items:
            type: string
        active:
          type: boolean
    NotificationTemplateInput:
      type: object
      required: [templateKey, templateName, templateType, htmlContent, textContent]
      properties:
        templateKey:
          type: string
        templateName:
          type: string
        templateType:
          type: string
          enum: [email, push, web_push]
        subject:
          type: string
        htmlContent:
          type: string
        textContent:
          type: string
        placeholders:
          type: array
          items:
            type: string
    WebPushSubscription:
      type: object
      required: [endpoint, keys]
      properties:
        endpoint:
          type: string
          format: uri
        keys:
          type: object
          required: [p256dh, auth]
          properties:
            p256dh:
              type: string
            auth:
              type: string
...